zheqi - move make_atoms_df and make_pairs_df to modules/dataframe_generation/dataframe_parse.py
